Don't wait too long
Using `networkidle0` waits until there is no remaining network activity
before taking the screenshot. This might be too conservative and we can
wait for no more than 2 with `networkidle2`, which seems to be the
practice for pages on the web.

This test works with `networkidle2` but throws a timeout error with
`networkidle0`:

```
node scripts/screenshot.js https://swap.stanford.edu/was/20171025171815/https://www.stanforddaily.com/2017/10/19/letter-to-the-editor-boards-response-to-su-prison-divest/ test.jpg
```

Closes #518