Set TMPDIR when sorting
When sorting CDXJ files `/tmp` will be used as temp space for sorting
when the files don't fit in memory. This is usually ok, but when the
CDXJ files being sorted are larger than what fits in `/tmp` an out of
space error will be encoutered. Setting `TMPDIR` to the space used by
CDXJ file generation uses should help solve this problem.

Closes #594