Allow PDF rendering for any URL
We have some fallback code to try a PDF render of a URL when headless
chrome fails to screenshot. Unfortunately that depends on the URL having
a `.pdf` at the end, which isn't really required for a PDF document on
the web.

This commit relaxes the constraint on the URL name, which will allow a
screenshot to be generated from a PDF at a URL like
https://wayback.archive-it.org/11377/20181120231857/http://sanjoseca.gov/DocumentCenter/View/79490/

Fixes #632