Automatically upgrade pipx and poetry
We ran into an issue during deployment where Poetry wasn't at the latest
version, and so we couldn't use the `package-mode` option in our
`pyproject.toml`.

This change will ensure we are running the latest versions of pipx and
poetry when doing a deployment. I guess the downside is that it could
get updated when doing the deployment, and lead to a failure that isn't
easy to observe in our dev environments?